Adding per-opcode execution count to debugging interpreter. Moved debugging interpreter guts to vm_debug source unit.
