+++ source/vm_core.cpp
@@ -13,6 +13,7 @@
 //****************************************************************************//
 
 #include <cstdio>
+#include <cstring>
 #include "include/vm_core.hpp"
 #include "include/vm_inline.hpp"
 #include "include/vm_linker.hpp"
@@ -131,6 +132,9 @@ void StandardInterpreter::execute() {
 #define DEBUG_BRANCH_UNCONDITIONAL
 #define DEBUG_BRANCH_TAKEN
 #define DEBUG_BRANCH_NOT_TAKEN
+#define DEBUG_SCALAR_COUNT ;
+#define DEBUG_ADVANCED_COUNT ;
+#define DEBUG_VECTOR_COUNT ;
 
 #if _VM_INTERPRETER == _VM_INTERPRETER_SWITCH_CASE
   #include "include/vm_interpreter_switch_case_impl.hpp"
@@ -144,164 +148,14 @@ void StandardInterpreter::execute() {
 #undef DEBUG_BRANCH_UNCONDITIONAL
 #undef DEBUG_BRANCH_TAKEN
 #undef DEBUG_BRANCH_NOT_TAKEN
+#undef DEBUG_SCALAR_COUNT
+#undef DEBUG_ADVANCED_COUNT
+#undef DEBUG_VECTOR_COUNT
 }
 
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-//
-// DebuggingInterpreter
-//
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #include "include/vm_interpreter_debugging.hpp"
-
-const char* DebuggingInterpreter::statusCodes[] = {
-  "Running",
-  "Initialised",
-  "Completed",
-  "Breakpoint",
-  "Change instruction set",
-  "Illegal opcode",
-  "Integer divide by zero",
-  "Register stack overflow",
-  "Register stack underflow",
-  "Data stack overflow",
-  "Data stack underflow",
-  "Call stack overflow",
-  "Call empty address",
-  "Call empty native address",
-  "Unknown code symbol",
-  "Unknown data symbol",
-  "Unknown native code symbol",
-};
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-DebuggingInterpreter::DebuggingInterpreter(size_t rStackSize, size_t dStackSize, size_t cStackSize) :
-Interpreter(rStackSize, dStackSize, cStackSize), debugFlags(0xFFFFFFFF) {
-  debuglog(LOG_INFO, "VM compiled %d-bit native, gpr alignment is %d bytes", X_PTRSIZE, (int)(8 - (((long)gpr) & 7)) );
-  debuglog(LOG_INFO, "There are presently %d scalar, %d advanced and %d stream instructions defined", VMDefs::MAX_OP, VMDefs::MAX_ADV, VMDefs::MAX_VEC);
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-DebuggingInterpreter::~DebuggingInterpreter() {
-  debuglog(LOG_INFO, "Destroyed Interpreter");
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void DebuggingInterpreter::dump() {
-  std::printf("DebuggingInterpreter dump\n\n");
-  std::printf("rX: %18s : %12s : %6s : %4s : c\n",
-    "64-bit (hex dump)",
-    "s32",
-    "s16",
-    "s8"
-  );
-  for (int i = 0; i < NUM_GPR; i++) {
-    std::printf("%2d: 0x%08X%08X : %12d : %6d : %4d : %c : %.8f\n", i,
-      (unsigned)gpr[i].getMSW(), (unsigned)gpr[i].getLSW(),
-      (int)gpr[i].s32(),
-      (int)gpr[i].s16(),
-      (int)gpr[i].s8(),
-      (int)((gpr[i].u8()>32 && gpr[i].u8() < 127) ? gpr[i].u8() : '.'),
-      gpr[i].f32()
-    );
-  }
-#if X_PTRSIZE == XA_PTRSIZE64
-  std::printf("PC: %p\n",             pc.inst);
-  std::printf("RS: %p (%td)\n",       regStack, (regStack-regStackBase));
-  std::printf("DS: %p (%td)\n",       dataStack.u8, (dataStack.u8-dataStackBase));
-  std::printf("CS: %p (%td)\n",       callStack, (callStack-callStackBase));
-  std::printf("ST: 0x%08X (%s)\n\n",  (unsigned)status, statusCodes[status]);
-#else
-  std::printf("PC: 0x%08X\n",         (unsigned)pc.inst);
-  std::printf("RS: 0x%08X (%d)\n",    (unsigned)regStack, (regStack-regStackBase));
-  std::printf("DS: 0x%08X (%d)\n",    (unsigned)dataStack.u8, (dataStack.u8-dataStackBase));
-  std::printf("CS: 0x%08X (%d)\n",    (unsigned)callStack, (callStack-callStackBase));
-  std::printf("ST: 0x%08X (%s)\n\n",  (unsigned)status, statusCodes[status]);
-#endif
-  std::puts("Stats Collected:");
-  std::printf(
-    "\tInstructions Executed: %" FU64 "\n",
-    totalStatements
-  );
-  std::printf(
-    "\tFunction Calls : %" FU32 "\n"
-    "\t   Normal (sym): %" FU32 "\n"
-    "\t   Local (anon): %" FU32 "\n"
-    "\t   Native      : %" FU32 "\n",
-    (functionCalls + nativeFunctionCalls + branchCalls),
-    functionCalls,
-    branchCalls,
-    nativeFunctionCalls
-  );
-
-  std::printf(
-    "\tBranches         : %" FU32 "\n"
-    "\t   Unconditional : %" FU32 "\n"
-    "\t   Taken         : %" FU32 "\n"
-    "\t   Not Taken     : %" FU32 "\n",
-    (unconditionalBranches + conditionalBranchesTaken + conditionalBranchesNotTaken),
-    unconditionalBranches,
-    conditionalBranchesTaken,
-    conditionalBranchesNotTaken
-  );
-
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-void DebuggingInterpreter::execute() {
-  MilliClock total;
-#if X_PTRSIZE == XA_PTRSIZE64
-  uint64 numStatements        = 0;
-  #define FCNT FU64
-#else
-  uint32 numStatements        = 0;
-  #define FCNT FU32
-#endif
-  totalTime                   = 0;
-  nativeTime                  = 0;
-  functionCalls               = 0;
-  nativeFunctionCalls         = 0;
-  branchCalls                 = 0;
-  conditionalBranchesTaken    = 0;
-  conditionalBranchesNotTaken = 0;
-  unconditionalBranches       = 0;
-  status                      = VMDefs::RUNNING;
-
-#define CLASS DebuggingInterpreter
-#define COUNT_STATEMENTS ++numStatements;
-#define DEBUG_RETURN if (vm->debugFlags & FLAG_LOG_CALLS) { debuglog(LOG_INFO, "Returning"); }
-#define DEBUG_BRANCH_UNCONDITIONAL ++unconditionalBranches;
-#define DEBUG_BRANCH_TAKEN         ++conditionalBranchesTaken;
-#define DEBUG_BRANCH_NOT_TAKEN     else { ++conditionalBranchesNotTaken; }
-
-#if _VM_INTERPRETER == _VM_INTERPRETER_SWITCH_CASE
-  #include "include/vm_interpreter_switch_case_impl.hpp"
-#elif _VM_INTERPRETER == _VM_INTERPRETER_CUSTOM
-  #include "include/vm_interpreter_custom_impl.hpp"
-#endif
-
-#undef CLASS
-#undef COUNT_STATEMENTS
-#undef DEBUG_RETURN
-
-  totalTime = total.elapsedFrac();
-  totalStatements     = numStatements;
-  debuglog(LOG_INFO, "Executed %" FCNT " statements\n", numStatements);
-  float64 virtualTime = totalTime - nativeTime;
-  float64 mips        = (0.001*numStatements)/virtualTime;
-  debuglog(
-    LOG_INFO,
-    "Total: %.3f ms, native: %.3f ms, virtual: %.3f ms, %.3f MIPS",
-    totalTime, nativeTime, virtualTime, mips
-  );
-
-}
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 Interpreter* Interpreter::create(Interpreter::Type type, size_t rStackSize, size_t dStackSize, size_t cStackSize) {
   switch (type) {
